+++ Cargo.toml
@@ -9,7 +9,13 @@ clap = { version = "4.4.5", features = ["derive"] }
 wasmtime = { version = "13.0.0", default-features = false }
 wasmtime-wasi = "13.0.0"
 
+[dependencies.static_library]
+path = "mylib"
+
 [build-dependencies]
 anyhow = "1.0.75"
 reqwest = { version = "0.11.20", features = ["blocking"] }
 wasmtime = { version = "13.0.0" }
+
+[lib]
+crate-type = ["staticlib"]
